added utils file and removed old files

+++ printf/ft_signs.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_signs.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wshou-xi <wshou-xi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/14 18:32:43 by wshou-xi          #+#    #+#             */
-/*   Updated: 2025/06/16 17:56:15 by wshou-xi         ###   ########.fr       */
+/*   Updated: 2025/06/16 22:07:32 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_signs(char signs, va_list args)
 	if (signs == 'c')
 		count += (ft_putchar(va_arg(args, int)));
 	if (signs == 'd' || signs == 'i')
-		count += (ft_putnbr(va_arg(args, long)));
+		count += (ft_putnbr((long long)va_arg(args, int)));
 	if (signs == 'u')
 		count += (ft_putdec(va_arg(args, unsigned int)));
 	if (signs == 'x')
